Update controllers, views and models

+++ controllers/usersController.php
@@ -10,7 +10,7 @@ class usersController extends Controller {
 
     public function __construct() {
         $u = new Users();
-        if (!$u->isLogged()) {
+        if(!$u->isLogged()) {
             header('location: '.BASE_URL.'/login');
         }
     }
@@ -19,20 +19,30 @@ class usersController extends Controller {
         $data = array();
         $u = new Users();
         $u->setLoggedUser();
-        $data['listUsers'] = $u->getList();
-        $data['nomeUser'] = $u->getEmail();
+        $data['listUsers'] = $u->getListUsers();
+        $data['nameUser'] = $u->getName();
         $this->loadTemplate('users', $data);
     }
 
     public function add() {
-        $u = new Users();
         $data = array();
+        $u = new Users();
+        $u->setLoggedUser();
+        $data['nameUser'] = $u->getName();
         if (isset($_POST['email']) && !empty($_POST['email'])) {
             $user = addslashes($_POST['name']);
             $email = addslashes($_POST['email']);
             $password = password_hash($_POST['password'], PASSWORD_DEFAULT);
+            $tipo_user = addslashes($_POST['tipo']);
+            $celular = addslashes($_POST['celular']);
 
-            $add = $u->add($user, $email, $password);
+            if(isset($_POST['nascimento']) && !empty($_POST['nascimento'])) {
+                $dt_nascimento = addslashes(date('Y-m-d',
+                                 strtotime(str_replace('/', '-', $_POST['nascimento']))));
+            } else {
+                $dt_nascimento = null;
+            }
+            $add = $u->add($user, $email, $password, $dt_nascimento, $tipo_user, $celular);
 
             if ($add) {
                 $data['feedback'] = 'Usuário cadastrado com suceeso!';
@@ -44,32 +54,44 @@ class usersController extends Controller {
     }
 
     public function edit($id) {
-        $u = new Users();
         $data = array();
-        $data['userData'] = $u->getUser($id);
+        $u = new Users();
+        $u->setLoggedUser();
+        $data['nameUser'] = $u->getName();
+        $data['userData'] = $u->getUserEdit($id); // obtém os dados do usuário a ser editado
+
         if (isset($_POST['name']) && !empty($_POST['name'])) {
             $name = addslashes($_POST['name']);
-            $u->edit($id, $name, '');
-            $data['feedback'] = 'Usuário Alterado!';
-        }
-        if (isset($_POST['password']) && !empty($_POST['password'])) {
-            if(isset($_POST['name']) && !empty($_POST['name'])) {
-                $pass = addslashes($_POST['password']);
-                $pass = password_hash($pass, PASSWORD_DEFAULT);
-                $u->edit($id, $name, $pass);
-                $data['feedback'] = 'Usuário Alterado!';
+            $pass = addslashes($_POST['password']);
+            $pass = password_hash($pass, PASSWORD_DEFAULT);
+            $tipo_user = addslashes($_POST['tipo']);
+            $status = addslashes($_POST['status']);
+
+            if(isset($_POST['nascimento']) && !empty($_POST['nascimento'])) {
+                $dt_nascimento = addslashes(date('Y-m-d',
+                                strtotime(str_replace('/', '-', $_POST['nascimento']))));
+            } else {
+                $dt_nascimento = null;
             }
+            if(isset($_POST['celular']) && !empty($_POST['celular'])) {
+                $celular = addslashes($_POST['celular']);
+            } else {
+                $celular = null;
+            }
+            $u->edit($id, $name, $pass, $dt_nascimento, $status, $tipo_user, $celular);
+            $data['feedback'] = 'Usuário Alterado com sucesso!';
         }
         $this->loadTemplate('user_edit', $data);
     }
 
     public function del($id) {
         $u = new Users();
-
-        if($u->delete($id)) {
-            echo 'Usuário Deletado!';
-        } else {
-            echo 'Não foi possível deleter o usuário!';
+        if ($u->isLogged()) {
+            if ($u->delete($id)) {
+                echo 'Usuário Deletado!';
+            } else {
+                echo 'Não foi possível deleter o usuário!';
+            }
         }
     }
 }
